Attempt to split scripts to avoid xvfb-run problems with python's OS library
